# docs/openapi/oauth-user-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Arkham Build OAuth and User API",
    "version": "1.0.0",
    "description": "OAuth 2.0 authorization-code API for confidential server-side clients. Client credentials are accepted only in form bodies. Native apps must forward authorization codes to a confidential backend; direct token exchange from an installed app is unsupported."
  },
...
  "servers": [
    {
      "url": "https://api.arkham.build"
    }
  ],
...
  "paths": {
...
    "/v2/user/me": {
      "get": {
        "operationId": "getOAuthUserProfile",
        "tags": ["User"],
        "summary": "Get the authenticated user's public profile",
        "description": "Requires profile:read. Email and identities are omitted.",
        "security": [
          {
            "OAuthBearer": ["profile:read"]
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated user's public profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthProfileResponse"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token is missing, malformed, expired, revoked, or otherwise unusable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            },
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer authentication challenge",
                "schema": {
                  "type": "string",
                  "examples": ["Bearer"]
                }
              }
            }
          },
          "403": {
            "description": "Account is banned or the token lacks profile:read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            },
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge, including the required scope when absent",
                "schema": {
                  "type": "string",
                  "examples": [
                    "Bearer error=\"insufficient_scope\", scope=\"decks:write\""
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v2/user/decks/manifest": {
      "get": {
        "operationId": "getOAuthDeckManifest",
        "tags": ["User decks"],
        "summary": "Get the user's deck manifest",
        "description": "Requires decks:read. Manifests that include ArkhamDB perform an upstream conditional sync and return an opaque snapshot token for consistent batch reads. ArkhamDB failure produces a partial 200 response with account decks and providers.arkhamdb.available=false.",
        "security": [
          {
            "OAuthBearer": ["decks:read"]
          }
        ],
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "description": "Optionally include only one provider",
            "schema": {
              "$ref": "#/components/schemas/OAuthDeckManifestQuery/properties/source"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Source-aware deck manifest, possibly partial when ArkhamDB is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthDeckManifestResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid source filter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token is unusable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            },
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer authentication challenge",
                "schema": {
                  "type": "string",
                  "examples": ["Bearer"]
                }
              }
            }
          },
          "403": {
            "description": "Token lacks decks:read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            },
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge, including the required scope when absent",
                "schema": {
                  "type": "string",
                  "examples": [
                    "Bearer error=\"insufficient_scope\", scope=\"decks:write\""
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v2/user/decks/batch": {
      "post": {
        "operationId": "getOAuthDeckBatch",
        "tags": ["User decks"],
        "summary": "Get a batch of decks",
        "description": "Requires decks:read. Accepts at most 250 provider-specific targets and preserves request order. ArkhamDB targets require the snapshot token returned by the manifest endpoint. The whole request fails if any target or the selected snapshot is unavailable.",
        "security": [
          {
            "OAuthBearer": ["decks:read"]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthDeckBatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Requested decks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthDeckBatchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Batch request is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token is unusable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            },
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer authentication challenge",
                "schema": {
                  "type": "string",
                  "examples": ["Bearer"]
                }
              }
            }
          },
          "403": {
            "description": "Token lacks decks:read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            },
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge, including the required scope when absent",
                "schema": {
                  "type": "string",
                  "examples": [
                    "Bearer error=\"insufficient_scope\", scope=\"decks:write\""
                  ]
                }
              }
            }
          },
          "404": {
            "description": "A deck was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            }
          },
          "409": {
            "description": "The selected ArkhamDB snapshot is no longer available",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            }
          },
          "503": {
            "description": "ArkhamDB is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            }
          }
        }
      }
    },
    "/v2/user/decks/{source}": {
      "post": {
        "operationId": "createOAuthDeck",
        "tags": ["User decks"],
        "summary": "Create a deck",
        "description": "Requires decks:write. Additional request properties, including server-owned deck fields, are ignored. Account creates generate a UUID; ArkhamDB chooses its ID.",
        "security": [
          {
            "OAuthBearer": ["decks:write"]
          }
        ],
        "parameters": [
          {
            "name": "source",
            "in": "path",
            "required": true,
            "description": "Deck provider: account or arkhamdb",
            "schema": {
              "$ref": "#/components/schemas/OAuthDeckSource"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthDeckWrite"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created deck",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthDeck"
                }
              }
            }
          },
          "400": {
            "description": "Deck input is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token is unusable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            },
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer authentication challenge",
                "schema": {
                  "type": "string",
                  "examples": ["Bearer"]
                }
              }
            }
          },
          "403": {
            "description": "Token lacks decks:write",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            },
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge, including the required scope when absent",
                "schema": {
                  "type": "string",
                  "examples": [
                    "Bearer error=\"insufficient_scope\", scope=\"decks:write\""
                  ]
                }
              }
            }
          },
          "503": {
            "description": "ArkhamDB is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            }
          }
        }
      }
    },
    "/v2/user/decks/{source}/{id}": {
      "get": {
        "operationId": "getOAuthDeck",
        "tags": ["User decks"],
        "summary": "Get one deck",
        "description": "Requires decks:read.",
        "security": [
          {
            "OAuthBearer": ["decks:read"]
          }
        ],
        "parameters": [
          {
            "name": "source",
            "in": "path",
            "required": true,
            "description": "Deck provider: account or arkhamdb",
            "schema": {
              "$ref": "#/components/schemas/OAuthDeckSource"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Account deck string ID or ArkhamDB numeric ID, encoded as a path segment",
            "schema": {
              "$ref": "#/components/schemas/OAuthDeckRouteId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Requested deck",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthDeck"
                }
              }
            }
          },
          "400": {
            "description": "Deck target is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token is unusable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            },
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer authentication challenge",
                "schema": {
                  "type": "string",
                  "examples": ["Bearer"]
                }
              }
            }
          },
          "403": {
            "description": "Token lacks decks:read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            },
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge, including the required scope when absent",
                "schema": {
                  "type": "string",
                  "examples": [
                    "Bearer error=\"insufficient_scope\", scope=\"decks:write\""
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Deck was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            }
          },
          "503": {
            "description": "ArkhamDB is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateOAuthDeck",
        "tags": ["User decks"],
        "summary": "Replace a deck",
        "description": "Requires decks:write. Fully replaces mutable content while preserving server-owned fields and history links.",
        "security": [
          {
            "OAuthBearer": ["decks:write"]
          }
        ],
        "parameters": [
          {
            "name": "source",
            "in": "path",
            "required": true,
            "description": "Deck provider: account or arkhamdb",
            "schema": {
              "$ref": "#/components/schemas/OAuthDeckSource"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Account deck string ID or ArkhamDB numeric ID, encoded as a path segment",
            "schema": {
              "$ref": "#/components/schemas/OAuthDeckRouteId"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthDeckWrite"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated deck",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthDeck"
                }
              }
            }
          },
          "400": {
            "description": "Deck input is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token is unusable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            },
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer authentication challenge",
                "schema": {
                  "type": "string",
                  "examples": ["Bearer"]
                }
              }
            }
          },
          "403": {
            "description": "Token lacks decks:write",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            },
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge, including the required scope when absent",
                "schema": {
                  "type": "string",
                  "examples": [
                    "Bearer error=\"insufficient_scope\", scope=\"decks:write\""
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Deck was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            }
          },
          "409": {
            "description": "Deck transition conflicts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            }
          },
          "503": {
            "description": "ArkhamDB is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteOAuthDeck",
        "tags": ["User decks"],
        "summary": "Delete a deck",
        "description": "Requires decks:delete. Set all=true to delete the selected deck and its previous history chain.",
        "security": [
          {
            "OAuthBearer": ["decks:delete"]
          }
        ],
        "parameters": [
          {
            "name": "source",
            "in": "path",
            "required": true,
            "description": "Deck provider: account or arkhamdb",
            "schema": {
              "$ref": "#/components/schemas/OAuthDeckSource"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Account deck string ID or ArkhamDB numeric ID, encoded as a path segment",
            "schema": {
              "$ref": "#/components/schemas/OAuthDeckRouteId"
            }
          },
          {
            "name": "all",
            "in": "query",
            "required": false,
            "description": "Delete the selected deck's previous history chain",
            "schema": {
              "$ref": "#/components/schemas/OAuthDeckDeleteQuery/properties/all"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deck deleted"
          },
          "400": {
            "description": "Deck target is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token is unusable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            },
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer authentication challenge",
                "schema": {
                  "type": "string",
                  "examples": ["Bearer"]
                }
              }
            }
          },
          "403": {
            "description": "Token lacks decks:delete",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            },
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge, including the required scope when absent",
                "schema": {
                  "type": "string",
                  "examples": [
                    "Bearer error=\"insufficient_scope\", scope=\"decks:write\""
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Deck was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            }
          },
          "503": {
            "description": "ArkhamDB is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            }
          }
        }
      }
    },
    "/v2/user/decks/{source}/{id}/upgrade": {
      "post": {
        "operationId": "upgradeOAuthDeck",
        "tags": ["User decks"],
        "summary": "Upgrade a deck",
        "description": "Requires decks:write. Creates a child deck and links it to the selected parent using provider-specific history semantics.",
        "security": [
          {
            "OAuthBearer": ["decks:write"]
          }
        ],
        "parameters": [
          {
            "name": "source",
            "in": "path",
            "required": true,
            "description": "Deck provider: account or arkhamdb",
            "schema": {
              "$ref": "#/components/schemas/OAuthDeckSource"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Account deck string ID or ArkhamDB numeric ID, encoded as a path segment",
            "schema": {
              "$ref": "#/components/schemas/OAuthDeckRouteId"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthDeckWrite"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created upgraded deck",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthDeck"
                }
              }
            }
          },
          "400": {
            "description": "Deck input is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token is unusable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            },
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer authentication challenge",
                "schema": {
                  "type": "string",
                  "examples": ["Bearer"]
                }
              }
            }
          },
          "403": {
            "description": "Token lacks decks:write",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            },
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge, including the required scope when absent",
                "schema": {
                  "type": "string",
                  "examples": [
                    "Bearer error=\"insufficient_scope\", scope=\"decks:write\""
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Deck was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            }
          },
          "409": {
            "description": "Deck already has an upgrade",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            }
          },
          "503": {
            "description": "ArkhamDB is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "OAuthDeckBatchRequest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "arkhamdbSyncToken": {
            "description": "Opaque token selecting the exact ArkhamDB snapshot returned by the manifest, or null when ArkhamDB was not included or available",
            "anyOf": [
              {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              {
                "type": "null"
              }
            ]
          },
          "decks": {
            "description": "At most 250 source-and-ID targets",
            "maxItems": 250,
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "type": "object",
                  "properties": {
                    "source": {
                      "type": "string",
                      "const": "account"
                    },
                    "id": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 255
                    }
                  },
                  "required": ["source", "id"],
                  "additionalProperties": false
                },
                {
                  "type": "object",
                  "properties": {
                    "source": {
                      "type": "string",
                      "const": "arkhamdb"
                    },
                    "id": {
                      "type": "integer",
                      "exclusiveMinimum": 0,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": ["source", "id"],
                  "additionalProperties": false
                }
              ]
            }
          }
        },
        "required": ["decks"],
        "additionalProperties": false
      },
...
      "OAuthProfileResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "id": {
            "description": "Stable arkham.build account ID",
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "username": {
            "type": "string"
          }
        },
        "required": ["id", "username"],
        "additionalProperties": false
      },
...
      "OAuthUserError": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "invalid_token",
              "insufficient_scope",
              "account_banned",
              "invalid_request",
              "not_found",
              "conflict",
              "upstream_unavailable"
            ]
          },
          "message": {
            "type": "string"
          }
        },
        "required": ["error", "message"],
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "OAuthBearer": {
        "type": "oauth2",
        "description": "Opaque access token sent as Authorization: Bearer ab_at_.... Tokens are issued only to confidential server-side clients.",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://api.arkham.build/v2/oauth/authorize",
            "tokenUrl": "https://api.arkham.build/v2/oauth/token",
            "scopes": {
              "profile:read": "Read the stable account ID and username",
              "decks:read": "Read account and connected ArkhamDB decks",
              "decks:write": "Create, replace, and upgrade decks; implies decks:read",
              "decks:delete": "Delete decks and history; implies decks:write and decks:read"
            }
          }
        }
      }
    }
  }
}